+++ lib/features/twitter/widgets/desktop/twitterMightLike/twitter_might_like_container.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+/// Contains the accounts to follow
+/// the title and the show more option
+class TwitterMightLikeContainer extends StatelessWidget {
+  const TwitterMightLikeContainer({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
