+++ lib/features/twitter/widgets/desktop/twitterMightLike/twitter_might_like_item.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+/// The accounts to follow
+class TwitterMightLikeItem extends StatelessWidget {
+  const TwitterMightLikeItem({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
