+++ lib/test/tweet_designs_main.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 
 import '../features/twitter/widgets/desktop/twitterMedia/twitter_media_container.dart';
 
-// TODO: Add the link text
 void main() => runApp(const MyApp());
 
 class MyApp extends StatelessWidget {
